add one shot multicalibrate
