+++ benchmarks/multivalid/mnist_top_label_multicalibrate.py
@@ -3,7 +3,10 @@ import numpy as np
 import tensorflow as tf
 import tensorflow_datasets as tfds
 
-from fortuna.conformal import TopLabelMulticalibrator
+from fortuna.conformal import (
+    OneShotTopLabelMulticalibrator,
+    TopLabelMulticalibrator,
+)
 from fortuna.data import DataLoader
 from fortuna.metric.classification import (
     accuracy,
@@ -79,56 +82,75 @@ for i in range(10):
 test_groups = np.stack(test_groups, axis=1)
 
 mc = TopLabelMulticalibrator(n_classes=10)
-calib_test_probs, status = mc.calibrate(
+mc_calib_test_probs, status = mc.calibrate(
     targets=val_targets,
     probs=val_means,
     groups=groups,
     test_probs=test_means,
     test_groups=test_groups,
-    n_buckets=1000,
 )
+mc_calib_val_probs = mc.apply_patches(probs=val_means, groups=groups)
 
-plt.plot(status["mean_squared_errors"], label="Mean-squared error decay")
-plt.legend()
-plt.xlabel("rounds")
-plt.show()
+osmc = OneShotTopLabelMulticalibrator(n_classes=10)
+osmc_calib_test_probs = osmc.calibrate(
+    targets=val_targets, probs=val_means, test_probs=test_means
+)
+osmc_calib_val_probs = osmc.apply_patches(probs=val_means)
 
-calib_val_probs = mc.apply_patches(probs=val_means, groups=groups)
 print(
     f"MSE on calibration data before calibration: {mc.mean_squared_error(val_means, val_targets)}"
 )
 print(
-    f"MSE on calibration data after calibration: {mc.mean_squared_error(calib_val_probs, val_targets)}"
+    f"MSE on calibration data after top-label multicalibration: {mc.mean_squared_error(mc_calib_val_probs, val_targets)}"
+)
+print(
+    f"MSE on calibration data after one-shot top-label multicalibration: {mc.mean_squared_error(osmc_calib_val_probs, val_targets)}"
 )
 print()
 print(
     f"MSE on test data before calibration: {mc.mean_squared_error(test_means, test_targets)}"
 )
 print(
-    f"MSE on test data after calibration: {mc.mean_squared_error(calib_test_probs, test_targets)}"
+    f"MSE on test data after top-label multicalibration: {mc.mean_squared_error(mc_calib_test_probs, test_targets)}"
+)
+print(
+    f"MSE on test data after one-shot top-label multicalibration: {osmc.mean_squared_error(osmc_calib_test_probs, test_targets)}"
 )
 print()
 print(
     f"ECE on calibration data before calibration: {expected_calibration_error(preds=val_modes, probs=val_means, targets=val_targets)}"
 )
 print(
-    f"ECE on calibration data after calibration: {expected_calibration_error(preds=calib_val_probs.argmax(1), probs=calib_val_probs, targets=val_targets)}"
+    f"ECE on calibration data after top-label multicalibration: {expected_calibration_error(preds=mc_calib_val_probs.argmax(1), probs=mc_calib_val_probs, targets=val_targets)}"
+)
+print(
+    f"ECE on calibration data after one-shot top-label multicalibration: {expected_calibration_error(preds=osmc_calib_val_probs.argmax(1), probs=osmc_calib_val_probs, targets=val_targets)}"
 )
 print()
 print(
     f"ECE on test data before calibration: {expected_calibration_error(preds=test_modes, probs=test_means, targets=test_targets)}"
 )
 print(
-    f"ECE on test data after calibration: {expected_calibration_error(preds=calib_test_probs.argmax(1), probs=calib_test_probs, targets=test_targets)}"
+    f"ECE on test data after top-label multicalibration: {expected_calibration_error(preds=mc_calib_test_probs.argmax(1), probs=mc_calib_test_probs, targets=test_targets)}"
+)
+print(
+    f"ECE on test data after one-shot top-label multicalibration: {expected_calibration_error(preds=osmc_calib_test_probs.argmax(1), probs=osmc_calib_test_probs, targets=test_targets)}"
 )
 print()
 print(
-    f"Accuracy on calibration data before calibration: {accuracy(val_modes, val_targets)}"
+    f"Accuracy on calibration data before top-label calibration: {accuracy(val_modes, val_targets)}"
 )
 print(
-    f"Accuracy on calibration data after calibration: {accuracy(calib_val_probs.argmax(1), val_targets)}"
+    f"Accuracy on calibration data after top-label multicalibration: {accuracy(mc_calib_val_probs.argmax(1), val_targets)}"
 )
+print(
+    f"Accuracy on calibration data after one-shot top-label multicalibration: {accuracy(osmc_calib_val_probs.argmax(1), val_targets)}"
+)
+print()
 print(f"Accuracy on test data before calibration: {accuracy(test_modes, test_targets)}")
 print(
-    f"Accuracy on test data after calibration: {accuracy(calib_test_probs.argmax(1), test_targets)}"
+    f"Accuracy on test data after top-label multicalibration: {accuracy(mc_calib_test_probs.argmax(1), test_targets)}"
+)
+print(
+    f"Accuracy on test data after one-shot top-label multicalibration: {accuracy(osmc_calib_test_probs.argmax(1), test_targets)}"
 )
