+++ fortuna/conformal/__init__.py
@@ -5,17 +5,25 @@ from fortuna.conformal.classification.adaptive_conformal_classifier import (
 from fortuna.conformal.classification.adaptive_prediction import (
     AdaptivePredictionConformalClassifier,
 )
-from fortuna.conformal.classification.batch_mvp import BatchMVPConformalClassifier
-from fortuna.conformal.classification.binary_multicalibrator import (
-    BinaryClassificationMulticalibrator,
-)
 from fortuna.conformal.classification.simple_prediction import (
     SimplePredictionConformalClassifier,
 )
-from fortuna.conformal.classification.top_label_multicalibrator import (
+from fortuna.conformal.multivalid.iterative.classification.binary_multicalibrator import (
+    BinaryClassificationMulticalibrator,
+)
+from fortuna.conformal.multivalid.iterative.classification.top_label_multicalibrator import (
     TopLabelMulticalibrator,
 )
-from fortuna.conformal.multivalid.multicalibrator import Multicalibrator
+from fortuna.conformal.multivalid.iterative.multicalibrator import Multicalibrator
+from fortuna.conformal.multivalid.iterative.regression.batch_mvp import (
+    BatchMVPConformalClassifier,
+)
+from fortuna.conformal.multivalid.one_shot.classification.binary_multicalibrator import (
+    OneShotBinaryClassificationMulticalibrator,
+)
+from fortuna.conformal.multivalid.one_shot.classification.top_label_multicalibrator import (
+    OneShotTopLabelMulticalibrator,
+)
 from fortuna.conformal.regression.adaptive_conformal_regressor import (
     AdaptiveConformalRegressor,
 )
