+++ fortuna/conformal/multivalid/iterative/classification/top_label_multicalibrator.py
@@ -13,11 +13,14 @@ from jax import (
 )
 import jax.numpy as jnp
 
-from fortuna.conformal.multivalid.multicalibrator import Multicalibrator
+from fortuna.conformal.multivalid.iterative.multicalibrator import Multicalibrator
+from fortuna.conformal.multivalid.mixins.classification.top_label_multicalibrator import (
+    TopLabelMulticalibratorMixin,
+)
 from fortuna.typing import Array
 
 
-class TopLabelMulticalibrator(Multicalibrator):
+class TopLabelMulticalibrator(TopLabelMulticalibratorMixin, Multicalibrator):
     def __init__(self, n_classes: int, seed: int = 0):
         """
         A multicalibration method that provides multivalid coverage guarantees. See Algorithm 15 in `Aaron Roth's notes
@@ -30,9 +33,7 @@ class TopLabelMulticalibrator(Multicalibrator):
         seed: int
             Random seed.
         """
-        super().__init__(seed=seed)
-        self._patch_list = []
-        self.n_classes = n_classes
+        super().__init__(n_classes, seed=seed)
 
     def calibrate(
         self,
@@ -43,7 +44,7 @@ class TopLabelMulticalibrator(Multicalibrator):
         test_probs: Optional[Array] = None,
         atol: float = 1e-4,
         rtol: float = 1e-6,
-        min_b_size: Union[float, int] = 0.1,
+        min_prob_b: float = 0.1,
         n_buckets: int = 100,
         n_rounds: int = 1000,
         eta: float = 0.1,
@@ -58,7 +59,7 @@ class TopLabelMulticalibrator(Multicalibrator):
             test_values=test_probs,
             atol=atol,
             rtol=rtol,
-            min_b_size=min_b_size,
+            min_prob_b=min_prob_b,
             n_buckets=n_buckets,
             n_rounds=n_rounds,
             eta=eta,
@@ -87,11 +88,6 @@ class TopLabelMulticalibrator(Multicalibrator):
             values=probs,
         )
 
-    def mean_squared_error(self, probs: Array, targets: Array) -> Array:
-        return super().mean_squared_error(
-            values=probs, scores=self._get_scores(targets)
-        )
-
     @staticmethod
     def _get_b(
         groups: Array, values: Array, v: Array, g: Array, c: Array, n_buckets: int
@@ -105,34 +101,8 @@ class TopLabelMulticalibrator(Multicalibrator):
     def _patch(
         self, values: Array, patch: Array, bt: Array, ct: Array, eta: float
     ) -> Array:
-        if jnp.all(~jnp.isnan(values)) and jnp.all(values.sum(1, keepdims=True) != 0.0):
-            values /= values.sum(1, keepdims=True)
+        values = self._maybe_normalize(values)
         return super()._patch(values=values, patch=patch, bt=bt, ct=ct, eta=eta)
-
-    @staticmethod
-    def _check_scores(scores: Array):
-        pass
-
-    @staticmethod
-    def _maybe_check_values(
-        values: Optional[Array], test_values: Optional[Array] = None
-    ):
-        if values is not None:
-            if values.ndim != 2:
-                raise ValueError(
-                    "`probs` must be a 2-dimensional array representing the probabilities for each input "
-                    "and each class."
-                )
-            if jnp.any(values < 0) or jnp.any(values > 1):
-                raise ValueError("All elements in `values` must be within [0, 1].")
-        if test_values is not None:
-            if test_values.ndim != 2:
-                raise ValueError(
-                    "`test_probs` must be a 2-dimensional array representing the probabilities for each "
-                    "input and each class."
-                )
-            if jnp.any(test_values < 0) or jnp.any(test_values > 1):
-                raise ValueError("All elements in `test_values` must be within [0, 1].")
 
     def _maybe_init_values(
         self, values: Optional[Array], size: Optional[int] = None
@@ -152,26 +122,3 @@ class TopLabelMulticalibrator(Multicalibrator):
         else:
             values = jnp.copy(values)
         return values
-
-    @staticmethod
-    def _round_to_buckets(v: Array, buckets: Array) -> Array:
-        def _fun(_v):
-            return buckets[jnp.argmin(jnp.abs(_v - buckets))]
-
-        if len(v.shape):
-            return vmap(_fun)(v)
-        return _fun(v)
-
-    def _get_scores(self, targets: Array) -> Array:
-        self._check_targets(targets)
-        scores = []
-        for i in range(self.n_classes):
-            scores.append(targets == i)
-        return jnp.stack(scores, axis=1)
-
-    @staticmethod
-    def _check_targets(targets: Array):
-        if targets.ndim != 1:
-            raise ValueError("`targets` must be a 1-dimensional array of integers.")
-        if targets.dtype not in ["int32", "int64"]:
-            raise ValueError("All elements in `targets` must be integers")
