+++ fortuna/conformal/multivalid/one_shot/base.py
@@ -7,17 +7,14 @@ from typing import (
     Union,
 )
 
-from jax import (
-    random,
-    vmap,
-)
+from jax import vmap
 import jax.numpy as jnp
-import numpy as np
 
+from fortuna.conformal.multivalid.base import MultivalidMethod
 from fortuna.typing import Array
 
 
-class MultivalidMethod:
+class OneShotMultivalidMethod(MultivalidMethod):
     def __init__(self, seed: int = 0):
         """
         A base one-shot multivalid method.
@@ -27,9 +24,9 @@ class MultivalidMethod:
         seed: int
             Random seed.
         """
-        self._patches = []
+        super().__init__(seed=seed)
+        self._patches = None
         self._n_buckets = None
-        self._seed = seed
 
     def calibrate(
         self,
@@ -37,7 +34,8 @@ class MultivalidMethod:
         values: Optional[Array] = None,
         test_values: Optional[Array] = None,
         n_buckets: int = 100,
-    ) -> Union[Dict, Tuple[Array, Dict]]:
+        min_prob_b: Union[float, str] = "auto",
+    ) -> Union[None, Array]:
         """
         Calibrate the model by finding a list of patches to the model that bring the calibration error below a
         certain threshold.
@@ -53,6 +51,9 @@ class MultivalidMethod:
             The initial model evaluations :math:`f(x)` on the test data. If not provided, these are set to 0.
         n_buckets: int
             The number of buckets used in the algorithm.
+        min_prob_b: Union[float, str]
+            Minimum probability of the conditioning set :math:`B_t` for the patch to be applied.
+            If "auto", it will be chosen based on the number of buckets and dimension of the scores.
 
         Returns
         -------
@@ -65,12 +66,19 @@ class MultivalidMethod:
             raise ValueError(
                 "If `test_values is provided, `values` must also be provided."
             )
+        if min_prob_b != "auto" and (min_prob_b < 0 or min_prob_b > 1):
+            raise ValueError(
+                "`min_prob_b` must be greater than or equal to 0 and less than or equal to 1."
+            )
 
         self._check_scores(scores)
         scores = self._process_scores(scores)
         n_dims = scores.shape[1]
 
-        values = self._maybe_init_values(values)
+        min_prob_b = self._maybe_init_min_prob_b(
+            min_prob_b=min_prob_b, n_buckets=n_buckets, n_dims=n_dims
+        )
+
         self._maybe_check_values(values, test_values)
 
         self.n_buckets = n_buckets
@@ -78,13 +86,68 @@ class MultivalidMethod:
 
         self._patches = vmap(
             lambda v: vmap(
-                lambda c: self._patch(
-                    v=v,
-                    c=c,
-                    scores=scores[:, c],
-                    values=values,
-                    n_buckets=n_buckets,
-                    **kwargs,
+                lambda c: self._get_patch(
+                    v=v, c=c, scores=scores[:, c], values=values, min_prob_b=min_prob_b
                 )
             )(jnp.arange(n_dims))
         )(buckets)
+
+        if test_values is not None:
+            return self.apply_patches(test_values)
+
+    def apply_patches(
+        self,
+        values: Array,
+    ) -> Array:
+        if self._patches is None:
+            logging.warning("No patches available.")
+            return values
+
+        buckets = self._get_buckets(self.n_buckets)
+        values = vmap(lambda v: self._round_to_buckets(v, buckets))(values)
+        unique_values = jnp.unique(values)
+
+        n_dims = 1 if values.ndim == 1 else values.shape[1]
+
+        b = vmap(
+            lambda v: vmap(
+                lambda c: self._get_b(values, v, c, self.n_buckets), out_axes=1
+            )(jnp.arange(n_dims)),
+            out_axes=1,
+        )(unique_values)
+
+        return self._patch(
+            values=values, b=b, unique_values=unique_values, buckets=buckets
+        )
+
+    @staticmethod
+    def _get_b(
+        values: Array,
+        v: Array,
+        c: Optional[Array],
+        n_buckets: int,
+    ) -> Array:
+        return jnp.abs(values - v) < 0.5 / n_buckets
+
+    def _patch(
+        self, values: Array, b: Array, unique_values: Array, buckets: Array
+    ) -> Array:
+        patched_values = jnp.copy(values)
+        for i, v in enumerate(unique_values):
+            for c in range(b.shape[2]):
+                idx_v = jnp.where(buckets == v)[0][0]
+                if values.ndim == 1:
+                    patched_values = patched_values.at[b[:, i, c]].set(
+                        self._patches[idx_v, c]
+                    )
+                else:
+                    patched_values = patched_values.at[b[:, i, c], c].set(
+                        self._patches[idx_v, c]
+                    )
+        return patched_values
+
+    @abc.abstractmethod
+    def _get_patch(
+        self, v: Array, c: Array, scores: Array, values: Array, min_prob_b: float
+    ) -> Array:
+        pass
